requirements.txt
acme==1.21.0
asgiref==3.7.2
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
httplib2==0.20.2
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
MarkupSafe==2.1.5
more-itertools==8.10.0
pillow==10.2.0
PyJWT==2.3.0
sqlparse==0.4.4
typing_extensions==4.9.0
